+++ database/gov.go
@@ -112,7 +112,7 @@ INSERT INTO proposal(
 		)
 	}
 
-	// Store the accounts
+	// Store the proposers accounts
 	err := db.SaveAccounts(accounts)
 	if err != nil {
 		return fmt.Errorf("error while storing proposers accounts: %s", err)
@@ -120,7 +120,17 @@ INSERT INTO proposal(
 
 	// Store the proposals
 	proposalsQuery = proposalsQuery[:len(proposalsQuery)-1] // Remove trailing ","
-	proposalsQuery += " ON CONFLICT DO NOTHING"
+	proposalsQuery += `
+ON CONFLICT (id) DO UPDATE 
+	SET title = excluded.title,
+  		description = excluded.description,
+		content = excluded.content,
+		proposer_address = excluded.proposer_address,
+		status = excluded.status,
+		submit_time = excluded.submit_time,
+		deposit_end_time = excluded.deposit_end_time,
+		voting_start_time = excluded.voting_start_time,
+		voting_end_time = excluded.voting_end_time`
 	_, err = db.SQL.Exec(proposalsQuery, proposalsParams...)
 	if err != nil {
 		return fmt.Errorf("error while storing proposals: %s", err)
@@ -134,7 +144,7 @@ func (db *Db) GetProposal(id uint64) (types.Proposal, error) {
 	var rows []*dbtypes.ProposalRow
 	err := db.SQL.Select(&rows, `SELECT * FROM proposal WHERE id = $1`, id)
 	if err != nil {
-		return types.Proposal{}, err
+		return types.Proposal{}, fmt.Errorf("error while getting proposal %d: %s", id, err)
 	}
 
 	if len(rows) == 0 {
@@ -203,7 +213,7 @@ func (db *Db) UpdateProposal(update types.ProposalUpdate) error {
 		update.ProposalID,
 	)
 	if err != nil {
-		return fmt.Errorf("error while updating proposal: %s", err)
+		return fmt.Errorf("error while updating proposal %d: %s", update.ProposalID, err)
 	}
 
 	return nil
@@ -215,25 +225,29 @@ func (db *Db) SaveDeposits(deposits []types.Deposit) error {
 		return nil
 	}
 
-	query := `INSERT INTO proposal_deposit (proposal_id, depositor_address, amount, timestamp, height) VALUES `
+	query := `INSERT INTO proposal_deposit (proposal_id, depositor_address, 
+		amount, timestamp, transaction_hash, height) VALUES `
 	var param []interface{}
 	var accounts []types.Account
 	for i, deposit := range deposits {
-		vi := i * 5
-		query += fmt.Sprintf("($%d,$%d,$%d,$%d,$%d),", vi+1, vi+2, vi+3, vi+4, vi+5)
+		vi := i * 6
+
+		accounts = append(accounts, types.NewAccount(deposit.Depositor))
+
+		query += fmt.Sprintf("($%d,$%d,$%d,$%d,$%d,$%d),", vi+1, vi+2, vi+3, vi+4, vi+5, vi+6)
 		param = append(param, deposit.ProposalID,
 			deposit.Depositor,
 			pq.Array(dbtypes.NewDbCoins(deposit.Amount)),
 			deposit.Timestamp,
+			deposit.TransactionHash,
 			deposit.Height,
 		)
-		accounts = append(accounts, types.NewAccount(deposit.Depositor))
 	}
 
-	// Store the depositor account
+	// Store depositors accounts
 	err := db.SaveAccounts(accounts)
 	if err != nil {
-		return fmt.Errorf("error while storing depositor account: %s", err)
+		return fmt.Errorf("error while storing depositors accounts: %s", err)
 	}
 
 	query = query[:len(query)-1] // Remove trailing ","
@@ -256,10 +270,10 @@ WHERE proposal_deposit.height <= excluded.height`
 // SaveVote allows to save for the given height and the message vote
 func (db *Db) SaveVote(vote types.Vote) error {
 	query := `
-INSERT INTO proposal_vote (proposal_id, voter_address, option, timestamp, height)
-VALUES ($1, $2, $3, $4, $5)
+INSERT INTO proposal_vote (proposal_id, voter_address, option, weight, timestamp, height)
+VALUES ($1, $2, $3, $4, $5, $6)
 ON CONFLICT ON CONSTRAINT unique_vote DO UPDATE
-	SET option = excluded.option,
+	SET weight = excluded.weight, 
 		timestamp = excluded.timestamp,
 		height = excluded.height
 WHERE proposal_vote.height <= excluded.height`
@@ -270,9 +284,9 @@ WHERE proposal_vote.height <= excluded.height`
 		return fmt.Errorf("error while storing voter account: %s", err)
 	}
 
-	_, err = db.SQL.Exec(query, vote.ProposalID, vote.Voter, vote.Option.String(), vote.Timestamp, vote.Height)
+	_, err = db.SQL.Exec(query, vote.ProposalID, vote.Voter, vote.Option.String(), vote.Weight, vote.Timestamp, vote.Height)
 	if err != nil {
-		return fmt.Errorf("error while storing vote: %s", err)
+		return fmt.Errorf("error while storing vote for proposal %d: %s", vote.ProposalID, err)
 	}
 
 	return nil
@@ -323,17 +337,19 @@ func (db *Db) SaveProposalStakingPoolSnapshot(snapshot types.ProposalStakingPool
 	stmt := `
 INSERT INTO proposal_staking_pool_snapshot (proposal_id, bonded_tokens, not_bonded_tokens, height)
 VALUES ($1, $2, $3, $4)
-ON CONFLICT ON CONSTRAINT unique_staking_pool_snapshot DO UPDATE SET
-	proposal_id = excluded.proposal_id,
-    bonded_tokens = excluded.bonded_tokens,
-	not_bonded_tokens = excluded.not_bonded_tokens,
+ON CONFLICT ON CONSTRAINT unique_staking_pool_snapshot DO UPDATE 
+	SET bonded_tokens = excluded.bonded_tokens,
+	not_bonded_tokens = excluded.not_bonded_tokens, 
 	height = excluded.height
 WHERE proposal_staking_pool_snapshot.height <= excluded.height`
 
 	_, err := db.SQL.Exec(stmt,
-		snapshot.ProposalID, snapshot.Pool.BondedTokens.String(), snapshot.Pool.NotBondedTokens.String(), snapshot.Pool.Height)
+		snapshot.ProposalID, snapshot.Pool.BondedTokens.String(),
+		snapshot.Pool.NotBondedTokens.String(), snapshot.Pool.Height)
+
 	if err != nil {
-		return fmt.Errorf("error while storing proposal staking pool snapshot: %s", err)
+		return fmt.Errorf("error while storing proposal staking pool snapshot for proposal %d: %s",
+			snapshot.ProposalID, err)
 	}
 
 	return nil
@@ -346,7 +362,7 @@ func (db *Db) SaveProposalValidatorsStatusesSnapshots(snapshots []types.Proposal
 	}
 
 	stmt := `
-INSERT INTO proposal_validator_status_snapshot(proposal_id, validator_address, voting_power, status, jailed, height)
+INSERT INTO proposal_validator_status_snapshot (proposal_id, validator_address, voting_power, status, jailed, height) 
 VALUES `
 
 	var args []interface{}
@@ -361,17 +377,15 @@ VALUES `
 
 	stmt = stmt[:len(stmt)-1]
 	stmt += `
-ON CONFLICT ON CONSTRAINT unique_validator_status_snapshot DO UPDATE
-	SET proposal_id = excluded.proposal_id,
-		validator_address = excluded.validator_address,
-		voting_power = excluded.voting_power,
-		status = excluded.status,
+ON CONFLICT ON CONSTRAINT unique_validator_status_snapshot DO UPDATE 
+	SET voting_power = excluded.voting_power, 
+		status = excluded.status, 
 		jailed = excluded.jailed,
 		height = excluded.height
 WHERE proposal_validator_status_snapshot.height <= excluded.height`
 	_, err := db.SQL.Exec(stmt, args...)
 	if err != nil {
-		return fmt.Errorf("error while storing proposal validator statuses snapshot: %s", err)
+		return fmt.Errorf("error while storing proposal validator statuses snapshots: %s", err)
 	}
 
 	return nil
@@ -393,7 +407,7 @@ WHERE software_upgrade_plan.height <= excluded.height`
 	_, err := db.SQL.Exec(stmt,
 		proposalID, plan.Name, plan.Height, plan.Info, height)
 	if err != nil {
-		return fmt.Errorf("error while storing software upgrade plan: %s", err)
+		return fmt.Errorf("error while storing software upgrade plan for proposal %d: %s", proposalID, err)
 	}
 
 	return nil
@@ -405,7 +419,7 @@ func (db *Db) DeleteSoftwareUpgradePlan(proposalID uint64) error {
 
 	_, err := db.SQL.Exec(stmt, proposalID)
 	if err != nil {
-		return fmt.Errorf("error while deleting software upgrade plan: %s", err)
+		return fmt.Errorf("error while deleting software upgrade plan for proposal %d: %s", proposalID, err)
 	}
 
 	return nil