+++ database/schema/08-gov.sql
@@ -28,8 +28,9 @@ CREATE TABLE proposal_deposit
     depositor_address TEXT REFERENCES account (address),
     amount            COIN[],
     timestamp         TIMESTAMP,
+    transaction_hash  TEXT    NOT NULL,
     height            BIGINT  NOT NULL,
-    CONSTRAINT unique_deposit UNIQUE (proposal_id, depositor_address)
+    CONSTRAINT unique_deposit UNIQUE (proposal_id, depositor_address, transaction_hash)
 );
 CREATE INDEX proposal_deposit_proposal_id_index ON proposal_deposit (proposal_id);
 CREATE INDEX proposal_deposit_depositor_address_index ON proposal_deposit (depositor_address);
@@ -40,9 +41,10 @@ CREATE TABLE proposal_vote
     proposal_id   INTEGER NOT NULL REFERENCES proposal (id),
     voter_address TEXT    NOT NULL REFERENCES account (address),
     option        TEXT    NOT NULL,
+    weight        TEXT    NOT NULL,
     timestamp     TIMESTAMP,
     height        BIGINT  NOT NULL,
-    CONSTRAINT unique_vote UNIQUE (proposal_id, voter_address)
+    CONSTRAINT unique_vote UNIQUE (proposal_id, voter_address, option)
 );
 CREATE INDEX proposal_vote_proposal_id_index ON proposal_vote (proposal_id);
 CREATE INDEX proposal_vote_voter_address_index ON proposal_vote (voter_address);
@@ -73,13 +75,13 @@ CREATE INDEX proposal_staking_pool_snapshot_proposal_id_index ON proposal_stakin
 
 CREATE TABLE proposal_validator_status_snapshot
 (
-    id                SERIAL PRIMARY KEY NOT NULL,
+    id                SERIAL  PRIMARY KEY NOT NULL,
     proposal_id       INTEGER REFERENCES proposal (id),
-    validator_address TEXT               NOT NULL REFERENCES validator (consensus_address),
-    voting_power      BIGINT             NOT NULL,
-    status            INT                NOT NULL,
-    jailed            BOOLEAN            NOT NULL,
-    height            BIGINT             NOT NULL,
+    validator_address TEXT                NOT NULL REFERENCES validator (consensus_address),
+    voting_power      BIGINT              NOT NULL,
+    status            INT                 NOT NULL,
+    jailed            BOOLEAN             NOT NULL,
+    height            BIGINT              NOT NULL,
     CONSTRAINT unique_validator_status_snapshot UNIQUE (proposal_id, validator_address)
 );
 CREATE INDEX proposal_validator_status_snapshot_proposal_id_index ON proposal_validator_status_snapshot (proposal_id);