+++ modules/gov/handle_msg.go
@@ -3,13 +3,16 @@ package gov
 import (
 	"fmt"
 	"strconv"
+	"strings"
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/x/authz"
 
 	"github.com/forbole/bdjuno/v4/types"
+	"google.golang.org/grpc/codes"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 
 	gov "github.com/cosmos/cosmos-sdk/x/gov/types"
@@ -36,12 +39,15 @@ func (m *Module) HandleMsg(index int, msg sdk.Msg, tx *juno.Tx) error {
 
 	case *govtypesv1.MsgVote:
 		return m.handleMsgVote(tx, cosmosMsg)
+
+	case *govtypesv1.MsgVoteWeighted:
+		return m.handleMsgVoteWeighted(tx, cosmosMsg)
 	}
 
 	return nil
 }
 
-// handleMsgSubmitProposal allows to properly handle a handleMsgSubmitProposal
+// handleMsgSubmitProposal allows to properly handle a MsgSubmitProposal
 func (m *Module) handleMsgSubmitProposal(tx *juno.Tx, index int, msg *govtypesv1.MsgSubmitProposal) error {
 	// Get the proposal id
 	event, err := tx.FindEventByType(index, gov.EventTypeSubmitProposal)
@@ -62,7 +68,44 @@ func (m *Module) handleMsgSubmitProposal(tx *juno.Tx, index int, msg *govtypesv1
 	// Get the proposal
 	proposal, err := m.source.Proposal(tx.Height, proposalID)
 	if err != nil {
-		return fmt.Errorf("error while getting proposal: %s", err)
+		if strings.Contains(err.Error(), codes.NotFound.String()) {
+			// query the proposal details using the latest height stored in db
+			// to fix the rpc error returning code = NotFound desc = proposal x doesn't exist
+			block, err := m.db.GetLastBlockHeightAndTimestamp()
+			if err != nil {
+				return fmt.Errorf("error while getting latest block height: %s", err)
+			}
+			proposal, err = m.source.Proposal(block.Height, proposalID)
+			if err != nil {
+				return fmt.Errorf("error while getting proposal: %s", err)
+			}
+		} else {
+			return fmt.Errorf("error while getting proposal: %s", err)
+		}
+	}
+
+	var addresses []types.Account
+	for _, msg := range proposal.Messages {
+		var sdkMsg sdk.Msg
+		err := m.cdc.UnpackAny(msg, &sdkMsg)
+		if err != nil {
+			return fmt.Errorf("error while unpacking proposal message: %s", err)
+		}
+
+		switch msg := sdkMsg.(type) {
+		case *distrtypes.MsgCommunityPoolSpend:
+			addresses = append(addresses, types.NewAccount(msg.Recipient))
+		case *govtypesv1.MsgExecLegacyContent:
+			content, ok := msg.Content.GetCachedValue().(*distrtypes.CommunityPoolSpendProposal)
+			if ok {
+				addresses = append(addresses, types.NewAccount(content.Recipient))
+			}
+		}
+	}
+
+	err = m.db.SaveAccounts(addresses)
+	if err != nil {
+		return fmt.Errorf("error while storing proposal recipient: %s", err)
 	}
 
 	// Store the proposal
@@ -91,11 +134,11 @@ func (m *Module) handleMsgSubmitProposal(tx *juno.Tx, index int, msg *govtypesv1
 	}
 
 	// Store the deposit
-	deposit := types.NewDeposit(proposal.Id, msg.Proposer, msg.InitialDeposit, txTimestamp, tx.Height)
+	deposit := types.NewDeposit(proposal.Id, msg.Proposer, msg.InitialDeposit, txTimestamp, tx.TxHash, tx.Height)
 	return m.db.SaveDeposits([]types.Deposit{deposit})
 }
 
-// handleMsgDeposit allows to properly handle a handleMsgDeposit
+// handleMsgDeposit allows to properly handle a MsgDeposit
 func (m *Module) handleMsgDeposit(tx *juno.Tx, msg *govtypesv1.MsgDeposit) error {
 	deposit, err := m.source.ProposalDeposit(tx.Height, msg.ProposalId, msg.Depositor)
 	if err != nil {
@@ -107,18 +150,43 @@ func (m *Module) handleMsgDeposit(tx *juno.Tx, msg *govtypesv1.MsgDeposit) error
 	}
 
 	return m.db.SaveDeposits([]types.Deposit{
-		types.NewDeposit(msg.ProposalId, msg.Depositor, deposit.Amount, txTimestamp, tx.Height),
+		types.NewDeposit(msg.ProposalId, msg.Depositor, deposit.Amount, txTimestamp, tx.TxHash, tx.Height),
 	})
 }
 
-// handleMsgVote allows to properly handle a handleMsgVote
+// handleMsgVote allows to properly handle a MsgVote
 func (m *Module) handleMsgVote(tx *juno.Tx, msg *govtypesv1.MsgVote) error {
 	txTimestamp, err := time.Parse(time.RFC3339, tx.Timestamp)
 	if err != nil {
 		return fmt.Errorf("error while parsing time: %s", err)
 	}
 
-	vote := types.NewVote(msg.ProposalId, msg.Voter, msg.Option, txTimestamp, tx.Height)
+	vote := types.NewVote(msg.ProposalId, msg.Voter, msg.Option, "1.0", txTimestamp, tx.Height)
 
-	return m.db.SaveVote(vote)
+	err = m.db.SaveVote(vote)
+	if err != nil {
+		return fmt.Errorf("error while saving vote: %s", err)
+	}
+
+	// update tally result for given proposal
+	return m.UpdateProposalTallyResult(msg.ProposalId, tx.Height)
+}
+
+// handleMsgVoteWeighted allows to properly handle a MsgVoteWeighted
+func (m *Module) handleMsgVoteWeighted(tx *juno.Tx, msg *govtypesv1.MsgVoteWeighted) error {
+	txTimestamp, err := time.Parse(time.RFC3339, tx.Timestamp)
+	if err != nil {
+		return fmt.Errorf("error while parsing time: %s", err)
+	}
+
+	for _, option := range msg.Options {
+		vote := types.NewVote(msg.ProposalId, msg.Voter, option.Option, option.Weight, txTimestamp, tx.Height)
+		err = m.db.SaveVote(vote)
+		if err != nil {
+			return fmt.Errorf("error while saving weighted vote for address %s: %s", msg.Voter, err)
+		}
+	}
+
+	// update tally result for given proposal
+	return m.UpdateProposalTallyResult(msg.ProposalId, tx.Height)
 }