+++ modules/gov/module.go
@@ -22,7 +22,6 @@ type Module struct {
 	cdc            codec.Codec
 	db             *database.Db
 	source         govsource.Source
-	authModule     AuthModule
 	distrModule    DistrModule
 	mintModule     MintModule
 	profilesModule ProfilesModule
@@ -33,7 +32,6 @@ type Module struct {
 // NewModule returns a new Module instance
 func NewModule(
 	source govsource.Source,
-	authModule AuthModule,
 	distrModule DistrModule,
 	mintModule MintModule,
 	profilesModule ProfilesModule,
@@ -45,7 +43,6 @@ func NewModule(
 	return &Module{
 		cdc:            cdc,
 		source:         source,
-		authModule:     authModule,
 		distrModule:    distrModule,
 		mintModule:     mintModule,
 		profilesModule: profilesModule,