+++ modules/gov/utils_proposal.go
@@ -3,13 +3,15 @@ package gov
 import (
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	govtypesv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	govtypesv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
 	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
+	"github.com/rs/zerolog/log"
+
+	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	proposaltypes "github.com/cosmos/cosmos-sdk/x/params/types/proposal"
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
@@ -18,16 +20,14 @@ import (
 
 	"google.golang.org/grpc/codes"
 
-	tmctypes "github.com/cometbft/cometbft/rpc/core/types"
-
 	"github.com/forbole/bdjuno/v4/types"
-
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 
 	gov "github.com/cosmos/cosmos-sdk/x/gov/types"
 )
 
-func (m *Module) UpdateProposal(height int64, blockTime time.Time, id uint64) error {
+// UpdateProposalStatus queries the latest details of given proposal ID, updates it's status
+// in database and handles changes if the proposal has been passed.
+func (m *Module) UpdateProposalStatus(height int64, id uint64) error {
 	// Get the proposal
 	proposal, err := m.source.Proposal(height, id)
 	if err != nil {
@@ -45,16 +45,6 @@ func (m *Module) UpdateProposal(height int64, blockTime time.Time, id uint64) er
 		return fmt.Errorf("error while updating proposal status: %s", err)
 	}
 
-	err = m.updateProposalTallyResult(proposal)
-	if err != nil {
-		return fmt.Errorf("error while updating proposal tally result: %s", err)
-	}
-
-	err = m.updateAccounts(proposal)
-	if err != nil {
-		return fmt.Errorf("error while updating account: %s", err)
-	}
-
 	err = m.handlePassedProposal(proposal, height)
 	if err != nil {
 		return fmt.Errorf("error while handling passed proposals: %s", err)
@@ -63,22 +53,53 @@ func (m *Module) UpdateProposal(height int64, blockTime time.Time, id uint64) er
 	return nil
 }
 
-func (m *Module) UpdateProposalValidatorStatusesSnapshot(height int64, blockVals *tmctypes.ResultValidators, id uint64) error {
-	err := m.updateProposalValidatorStatusesSnapshot(height, id, blockVals)
+// updateProposalStatus updates given proposal status
+func (m *Module) updateProposalStatus(proposal *govtypesv1.Proposal) error {
+	return m.db.UpdateProposal(
+		types.NewProposalUpdate(
+			proposal.Id,
+			proposal.Status.String(),
+			proposal.VotingStartTime,
+			proposal.VotingEndTime,
+		),
+	)
+}
+
+// UpdateProposalsStakingPoolSnapshot updates
+// staking pool snapshots for active proposals
+func (m *Module) UpdateProposalsStakingPoolSnapshot() error {
+	log.Debug().Str("module", "gov").Msg("refreshing proposal staking pool snapshots")
+	block, err := m.db.GetLastBlockHeightAndTimestamp()
 	if err != nil {
-		return fmt.Errorf("error while updating proposal validator statuses snapshot: %s", err)
+		return err
+	}
+
+	ids, err := m.db.GetOpenProposalsIds(block.BlockTimestamp)
+	if err != nil {
+		log.Error().Err(err).Str("module", "gov").Msg("error while getting open proposals ids")
+	}
+
+	for _, proposalID := range ids {
+		err = m.UpdateProposalStakingPoolSnapshot(block.Height, proposalID)
+		if err != nil {
+			return fmt.Errorf("error while updating proposal %d staking pool snapshots: %s", proposalID, err)
+		}
 	}
 
 	return nil
 }
 
-func (m *Module) UpdateProposalStakingPoolSnapshot(height int64, blockVals *tmctypes.ResultValidators, id uint64) error {
-	err := m.updateProposalStakingPoolSnapshot(height, id)
+// UpdateProposalStakingPoolSnapshot updates the staking pool snapshot associated with the gov
+// proposal having the provided id
+func (m *Module) UpdateProposalStakingPoolSnapshot(height int64, proposalID uint64) error {
+	pool, err := m.stakingModule.GetStakingPoolSnapshot(height)
 	if err != nil {
-		return fmt.Errorf("error while updating proposal staking pool snapshot: %s", err)
+		return fmt.Errorf("error while getting staking pool: %s", err)
 	}
 
-	return nil
+	return m.db.SaveProposalStakingPoolSnapshot(
+		types.NewProposalStakingPoolSnapshot(proposalID, pool),
+	)
 }
 
 // updateDeletedProposalStatus updates the proposal having the given id by setting its status
@@ -143,33 +164,39 @@ func (m *Module) handleParamChangeProposal(height int64, moduleName string) (err
 	return nil
 }
 
-// updateProposalStatus updates the given proposal status
-func (m *Module) updateProposalStatus(proposal *govtypesv1.Proposal) error {
-	return m.db.UpdateProposal(
-		types.NewProposalUpdate(
-			proposal.Id,
-			proposal.Status.String(),
-			proposal.VotingStartTime,
-			proposal.VotingEndTime,
-		),
-	)
-}
-
-// updateProposalTallyResult updates the tally result associated with the given proposal
-func (m *Module) updateProposalTallyResult(proposal *govtypesv1.Proposal) error {
-	height, err := m.db.GetLastBlockHeight()
+// UpdateProposalsTallyResults updates the tally for active proposals
+func (m *Module) UpdateProposalsTallyResults() error {
+	log.Debug().Str("module", "gov").Msg("refreshing proposal tally results")
+	block, err := m.db.GetLastBlockHeightAndTimestamp()
 	if err != nil {
 		return err
 	}
 
-	result, err := m.source.TallyResult(height, proposal.Id)
+	ids, err := m.db.GetOpenProposalsIds(block.BlockTimestamp)
+	if err != nil {
+		log.Error().Err(err).Str("module", "gov").Msg("error while getting open proposals ids")
+	}
+
+	for _, proposalID := range ids {
+		err = m.UpdateProposalTallyResult(proposalID, block.Height)
+		if err != nil {
+			return fmt.Errorf("error while updating proposal %d tally result : %s", proposalID, err)
+		}
+	}
+
+	return nil
+}
+
+// UpdateProposalTallyResult updates the tally result associated with the given proposal ID
+func (m *Module) UpdateProposalTallyResult(proposalID uint64, height int64) error {
+	result, err := m.source.TallyResult(height, proposalID)
 	if err != nil {
 		return fmt.Errorf("error while getting tally result: %s", err)
 	}
 
 	return m.db.SaveTallyResults([]types.TallyResult{
 		types.NewTallyResult(
-			proposal.Id,
+			proposalID,
 			result.YesCount,
 			result.AbstainCount,
 			result.NoCount,
@@ -179,118 +206,6 @@ func (m *Module) updateProposalTallyResult(proposal *govtypesv1.Proposal) error 
 	})
 }
 
-// updateAccounts updates any account that might be involved in the proposal (eg. fund community recipient)
-func (m *Module) updateAccounts(proposal *govtypesv1.Proposal) error {
-	var addresses []string
-	for _, msg := range proposal.Messages {
-		var sdkMsg sdk.Msg
-		err := m.cdc.UnpackAny(msg, &sdkMsg)
-		if err != nil {
-			return fmt.Errorf("error while unpacking proposal message: %s", err)
-		}
-
-		switch msg := sdkMsg.(type) {
-		case *distrtypes.MsgCommunityPoolSpend:
-			addresses = append(addresses, msg.Recipient)
-		case *govtypesv1.MsgExecLegacyContent:
-			content, ok := msg.Content.GetCachedValue().(*distrtypes.CommunityPoolSpendProposal)
-			if ok {
-				addresses = append(addresses, content.Recipient)
-			}
-		}
-	}
-
-	height, err := m.db.GetLastBlockHeight()
-	if err != nil {
-		return fmt.Errorf("error while getting last block height: %s", err)
-	}
-
-	return m.authModule.RefreshAccounts(height, addresses)
-}
-
-// updateProposalStakingPoolSnapshot updates the staking pool snapshot associated with the gov
-// proposal having the provided id
-func (m *Module) updateProposalStakingPoolSnapshot(height int64, proposalID uint64) error {
-	pool, err := m.stakingModule.GetStakingPoolSnapshot(height)
-	if err != nil {
-		return fmt.Errorf("error while getting staking pool: %s", err)
-	}
-
-	return m.db.SaveProposalStakingPoolSnapshot(
-		types.NewProposalStakingPoolSnapshot(proposalID, pool),
-	)
-}
-
-// updateProposalValidatorStatusesSnapshot updates the snapshots of the various validators for
-// the proposal having the given id
-func (m *Module) updateProposalValidatorStatusesSnapshot(
-	height int64, proposalID uint64, blockVals *tmctypes.ResultValidators,
-) error {
-	validators, _, err := m.stakingModule.GetValidatorsWithStatus(height, stakingtypes.Bonded.String())
-	if err != nil {
-		return fmt.Errorf("error while getting validators with bonded status: %s", err)
-	}
-
-	votingPowers, err := m.stakingModule.GetValidatorsVotingPowers(height, blockVals)
-	if err != nil {
-		return fmt.Errorf("error while getting validators voting powers: %s", err)
-	}
-
-	statuses, err := m.stakingModule.GetValidatorsStatuses(height, validators)
-	if err != nil {
-		return fmt.Errorf("error while getting validator statuses: %s", err)
-	}
-
-	var snapshots = make([]types.ProposalValidatorStatusSnapshot, len(validators))
-	for index, validator := range validators {
-		consAddr, err := validator.GetConsAddr()
-		if err != nil {
-			return err
-		}
-
-		status, err := findStatus(consAddr.String(), statuses)
-		if err != nil {
-			return fmt.Errorf("error while searching for status: %s", err)
-		}
-
-		votingPower, err := findVotingPower(consAddr.String(), votingPowers)
-		if err != nil {
-			return fmt.Errorf("error while searching for voting power: %s", err)
-		}
-
-		snapshots[index] = types.NewProposalValidatorStatusSnapshot(
-			proposalID,
-			consAddr.String(),
-			votingPower.VotingPower,
-			status.Status,
-			status.Jailed,
-			height,
-		)
-	}
-
-	return m.db.SaveProposalValidatorsStatusesSnapshots(snapshots)
-}
-
-func findVotingPower(consAddr string, powers []types.ValidatorVotingPower) (types.ValidatorVotingPower, error) {
-	for _, votingPower := range powers {
-		if votingPower.ConsensusAddress == consAddr {
-			return votingPower, nil
-		}
-	}
-	return types.ValidatorVotingPower{}, fmt.Errorf("voting power not found for validator with consensus address %s", consAddr)
-}
-
-func findStatus(consAddr string, statuses []types.ValidatorStatus) (types.ValidatorStatus, error) {
-	for _, status := range statuses {
-		if status.ConsensusAddress == consAddr {
-			return status, nil
-		}
-	}
-	return types.ValidatorStatus{}, fmt.Errorf("cannot find status for validator with consensus address %s", consAddr)
-}
-
-// handlePassedProposal handles a passed proposal by updating the proposal status and
-// updating any other data that might be involved in the proposal (eg. fund community recipient)
 func (m *Module) handlePassedProposal(proposal *govtypesv1.Proposal, height int64) error {
 	if proposal.Status != govtypesv1.StatusPassed {
 		// If proposal status is not passed, do nothing