+++ modules/staking/handle_block.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/forbole/bdjuno/v4/types"
 
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	juno "github.com/forbole/juno/v5/types"
 
 	tmctypes "github.com/cometbft/cometbft/rpc/core/types"
@@ -19,16 +18,10 @@ func (m *Module) HandleBlock(
 	block *tmctypes.ResultBlock, res *tmctypes.ResultBlockResults, _ []*juno.Tx, vals *tmctypes.ResultValidators,
 ) error {
 	// Update the validators
-	validators, err := m.updateValidators(block.Block.Height)
+	_, err := m.updateValidators(block.Block.Height)
 	if err != nil {
 		return fmt.Errorf("error while updating validators: %s", err)
 	}
-
-	// Update the voting powers
-	go m.updateValidatorVotingPower(block.Block.Height, vals)
-
-	// Update the validators statuses
-	go m.updateValidatorsStatus(block.Block.Height, validators)
 
 	// Updated the double sign evidences
 	go m.updateDoubleSignEvidence(block.Block.Height, block.Block.Evidence.Evidence)
@@ -36,60 +29,19 @@ func (m *Module) HandleBlock(
 	return nil
 }
 
-// updateValidatorsStatus updates all validators' statuses
-func (m *Module) updateValidatorsStatus(height int64, validators []stakingtypes.Validator) {
-	log.Debug().Str("module", "staking").Int64("height", height).
-		Msg("updating validators statuses")
-
-	statuses, err := m.GetValidatorsStatuses(height, validators)
-	if err != nil {
-		log.Error().Str("module", "staking").Err(err).
-			Int64("height", height).
-			Send()
-		return
-	}
-
-	err = m.db.SaveValidatorsStatuses(statuses)
-	if err != nil {
-		log.Error().Str("module", "staking").Err(err).
-			Int64("height", height).
-			Msg("error while saving validators statuses")
-	}
-}
-
-// updateValidatorVotingPower fetches and stores into the database all the current validators' voting powers
-func (m *Module) updateValidatorVotingPower(height int64, vals *tmctypes.ResultValidators) {
-	log.Debug().Str("module", "staking").Int64("height", height).
-		Msg("updating validators voting powers")
-
-	// Get the voting powers
-	votingPowers, err := m.GetValidatorsVotingPowers(height, vals)
-	if err != nil {
-		log.Error().Str("module", "staking").Err(err).Int64("height", height).
-			Msg("error while getting validators voting powers")
-		return
-	}
-
-	// Save all the voting powers
-	err = m.db.SaveValidatorsVotingPowers(votingPowers)
-	if err != nil {
-		log.Error().Str("module", "staking").Err(err).Int64("height", height).
-			Msg("error while saving validators voting powers")
-	}
-}
-
 // updateDoubleSignEvidence updates the double sign evidence of all validators
 func (m *Module) updateDoubleSignEvidence(height int64, evidenceList tmtypes.EvidenceList) {
 	log.Debug().Str("module", "staking").Int64("height", height).
 		Msg("updating double sign evidence")
 
+	var evidences []types.DoubleSignEvidence
 	for _, ev := range evidenceList {
 		dve, ok := ev.(*tmtypes.DuplicateVoteEvidence)
 		if !ok {
 			continue
 		}
 
-		evidence := types.NewDoubleSignEvidence(
+		evidences = append(evidences, types.NewDoubleSignEvidence(
 			height,
 			types.NewDoubleSignVote(
 				int(dve.VoteA.Type),
@@ -109,14 +61,15 @@ func (m *Module) updateDoubleSignEvidence(height int64, evidenceList tmtypes.Evi
 				dve.VoteB.ValidatorIndex,
 				hex.EncodeToString(dve.VoteB.Signature),
 			),
+		),
 		)
-
-		err := m.db.SaveDoubleSignEvidence(evidence)
-		if err != nil {
-			log.Error().Str("module", "staking").Err(err).Int64("height", height).
-				Msg("error while saving double sign evidence")
-			return
-		}
-
 	}
+
+	err := m.db.SaveDoubleSignEvidences(evidences)
+	if err != nil {
+		log.Error().Str("module", "staking").Err(err).Int64("height", height).
+			Msg("error while saving double sign evidence")
+		return
+	}
+
 }