+++ modules/staking/handle_msg.go
@@ -27,6 +27,19 @@ func (m *Module) HandleMsg(_ int, msg sdk.Msg, tx *juno.Tx) error {
 
 	case *stakingtypes.MsgEditValidator:
 		return m.handleEditValidator(tx.Height, cosmosMsg)
+
+	// update validators statuses, voting power
+	// and proposals validators satatus snapshots
+	// when there is a voting power change
+	case *stakingtypes.MsgDelegate:
+		return m.UpdateValidatorStatuses()
+
+	case *stakingtypes.MsgBeginRedelegate:
+		return m.UpdateValidatorStatuses()
+
+	case *stakingtypes.MsgUndelegate:
+		return m.UpdateValidatorStatuses()
+
 	}
 
 	return nil