+++ modules/staking/handle_periodic_operations.go
@@ -20,19 +20,26 @@ func (m *Module) RegisterPeriodicOperations(scheduler *gocron.Scheduler) error {
 		return fmt.Errorf("error while scheduling staking pool periodic operation: %s", err)
 	}
 
+	// refresh proposal validators status snapshots every 5 mins
+	if _, err := scheduler.Every(5).Minutes().Do(func() {
+		utils.WatchMethod(m.UpdateValidatorStatuses)
+	}); err != nil {
+		return fmt.Errorf("error while setting up gov period operations: %s", err)
+	}
+
 	return nil
 }
 
 // UpdateStakingPool reads from the LCD the current staking pool and stores its value inside the database
 func (m *Module) UpdateStakingPool() error {
-	height, err := m.db.GetLastBlockHeight()
+	block, err := m.db.GetLastBlockHeightAndTimestamp()
 	if err != nil {
 		return fmt.Errorf("error while getting latest block height: %s", err)
 	}
-	log.Debug().Str("module", "staking").Int64("height", height).
+	log.Debug().Str("module", "staking").Int64("height", block.Height).
 		Msg("updating staking pool")
 
-	pool, err := m.GetStakingPool(height)
+	pool, err := m.GetStakingPool(block.Height)
 	if err != nil {
 		return fmt.Errorf("error while getting staking pool: %s", err)