+++ modules/staking/utils_validators.go
@@ -3,9 +3,6 @@ package staking
 import (
 	"fmt"
 
-	tmctypes "github.com/cometbft/cometbft/rpc/core/types"
-	juno "github.com/forbole/juno/v5/types"
-
 	"github.com/forbole/bdjuno/v4/modules/staking/keybase"
 	"github.com/forbole/bdjuno/v4/types"
 
@@ -204,35 +201,120 @@ func (m *Module) GetValidatorsStatuses(height int64, validators []stakingtypes.V
 	return statuses, nil
 }
 
-func (m *Module) GetValidatorsVotingPowers(height int64, vals *tmctypes.ResultValidators) ([]types.ValidatorVotingPower, error) {
-	stakingVals, _, err := m.getValidators(height)
+// UpdateValidatorStatuses allows to update validators status, voting power
+// and active proposals validator status snapshots
+func (m *Module) UpdateValidatorStatuses() error {
+	// get the latest block height from db
+	block, err := m.db.GetLastBlockHeightAndTimestamp()
 	if err != nil {
-		return nil, err
+		return fmt.Errorf("error while getting latest block height from db: %s", err)
 	}
 
-	votingPowers := make([]types.ValidatorVotingPower, len(stakingVals))
-	for index, validator := range stakingVals {
-		// Get the validator consensus address
+	validators, _, err := m.GetValidatorsWithStatus(block.Height, stakingtypes.Bonded.String())
+	if err != nil {
+		return fmt.Errorf("error while getting validators with bonded status: %s", err)
+	}
+
+	// update validators status and voting power in database
+	err = m.updateValidatorStatusAndVP(block.Height, validators)
+	if err != nil {
+		return fmt.Errorf("error while updating validators status and voting power: %s", err)
+	}
+
+	// get all active proposals IDs from db
+	ids, err := m.db.GetOpenProposalsIds(block.BlockTimestamp)
+	if err != nil {
+		return fmt.Errorf("error while getting open proposals ids: %s", err)
+	}
+
+	// update validator status snapshots for all proposals IDs
+	// returned from database
+	for _, id := range ids {
+		// update validator status snapshot for given height and proposal ID
+		err = m.updateProposalValidatorStatusSnapshot(block.Height, id, validators)
+		if err != nil {
+			return fmt.Errorf("error while updating proposal validator status snapshots: %s", err)
+		}
+	}
+
+	return nil
+}
+
+// updateProposalValidatorStatusSnapshot updates validators snapshot for
+// the proposal having the given id
+func (m *Module) updateProposalValidatorStatusSnapshot(
+	height int64, proposalID uint64, validators []stakingtypes.Validator) error {
+	snapshots := make([]types.ProposalValidatorStatusSnapshot, len(validators))
+
+	for index, validator := range validators {
 		consAddr, err := validator.GetConsAddr()
 		if err != nil {
-			return nil, err
+			return err
 		}
 
-		// Find the voting power of this validator
-		var votingPower int64 = 0
-		for _, blockVal := range vals.Validators {
-			blockValConsAddr := juno.ConvertValidatorAddressToBech32String(blockVal.Address)
-			if blockValConsAddr == consAddr.String() {
-				votingPower = blockVal.VotingPower
-			}
+		snapshots[index] = types.NewProposalValidatorStatusSnapshot(
+			proposalID,
+			consAddr.String(),
+			validator.Tokens.Int64(),
+			validator.Status,
+			validator.Jailed,
+			height,
+		)
+	}
+
+	log.Debug().Str("module", "staking").Msg("refreshing proposal validator statuses snapshots")
+
+	return m.db.SaveProposalValidatorsStatusesSnapshots(snapshots)
+}
+
+// updateValidatorStatusAndVP updates validators status
+// and validators voting power
+func (m *Module) updateValidatorStatusAndVP(height int64, validators []stakingtypes.Validator) error {
+	votingPowers := make([]types.ValidatorVotingPower, len(validators))
+	statuses := make([]types.ValidatorStatus, len(validators))
+
+	for index, validator := range validators {
+		consAddr, err := validator.GetConsAddr()
+		if err != nil {
+			return err
 		}
 
 		if found, _ := m.db.HasValidator(consAddr.String()); !found {
 			continue
 		}
 
-		votingPowers[index] = types.NewValidatorVotingPower(consAddr.String(), votingPower, height)
+		consPubKey, err := m.getValidatorConsPubKey(validator)
+		if err != nil {
+			return err
+		}
+
+		votingPowers[index] = types.NewValidatorVotingPower(consAddr.String(), validator.Tokens.Int64(), height)
+
+		statuses[index] = types.NewValidatorStatus(
+			consAddr.String(),
+			consPubKey.String(),
+			int(validator.GetStatus()),
+			validator.IsJailed(),
+			height,
+		)
 	}
 
-	return votingPowers, nil
+	log.Debug().Str("module", "staking").Msg("refreshing validator voting power")
+	// Save validators voting powers in db
+	err := m.db.SaveValidatorsVotingPowers(votingPowers)
+	if err != nil {
+		log.Error().Str("module", "staking").Err(err).Int64("height", height).
+			Msg("error while saving validators voting powers")
+	}
+
+	log.Debug().Str("module", "staking").Msg("refreshing validator statuses")
+	// Save validators statuses in db
+	err = m.db.SaveValidatorsStatuses(statuses)
+	if err != nil {
+		log.Error().Str("module", "staking").Err(err).
+			Int64("height", height).
+			Msg("error while saving validators statuses")
+	}
+
+	return nil
 }